\name{resortDeaths}
\alias{resortDeaths}
\title{\code{resortDeaths()} utility to sort Death inputs by Lexis within Sex within Age within Year}
\usage{
  resortDeaths(deathsInHMDFormat)
}
\arguments{
  \item{deathsInHMDFormat}{data.frame}
}
\value{
  deathsInHMDFormat data.frame, resorted
}
\description{
  does as promised
}

