\name{resortLDB}
\alias{resortLDB}
\title{\code{resortLDB()} utility to sort LexisDB objects by Lexis within Age within Year}
\usage{
  resortLDB(x)
}
\arguments{
  \item{Death}{data.frame}
}
\value{
  Death data.frame, resorted
}
\description{
  LexisDB objects should actually already be sorted, and
  this function was written just for the sake of comparing
  output while getting a handle on the matlab scripts.
  Columns must already be in standard order.
}

