\name{resortMonthly}
\alias{resortMonthly}
\title{\code{resortMonthly()} utility to sort Death inputs by Month within Year}
\usage{
  resortMonthly(XXXmonthly)
}
\arguments{
  \item{XXXmonthly}{data.frame}
}
\value{
  XXXmonthly data.frame, resorted
}
\description{
  This data input has no Sex column at present.
}

