\name{resortPop}
\alias{resortPop}
\title{\code{resortPop()} utility to sort Population inputs by Age within Sex within Month within Year}
\usage{
  resortPop(popInHMDFormat)
}
\arguments{
  \item{popInHMDFormat}{data.frame}
}
\value{
  popInHMDFormat data.frame, resorted
}
\description{
  does as promised
}

